requirements.txt
alembic
asn1crypto
Babel
bcrypt
blinker
certifi
cffi
chardet
click
coverage
cryptography
dominate
elasticsearch
Flask
Flask-Babel
Flask-Bootstrap
Flask-HTTPAuth
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-SQLAlchemy
Flask-WTF
guess-language-spirit
idna
itsdangerous
Jinja2
Mako
MarkupSafe
passlib
pyasn1
pycparser
PyJWT
python-dateutil
python-dotenv
python-editor
pytz
PyYAML
redis
requests
rq
rsa
six
SQLAlchemy
SQLAlchemy-Utils
ujson
urllib3
visitor
Werkzeug
WTForms
pyqrcode
onetimepass
flask-bcrypt